feat: set height limit on gist and extract gist to separate component
